+++ src/features/definitionProvider.ts
@@ -15,8 +15,6 @@ import LogtalkTerminal from "./terminal";
 import { getLogger } from "../utils/logger";
 import { Utils } from "../utils/utils";
 import * as path from "path";
-import * as fs from "fs";
-import * as fsp from "fs/promises";
 
 export class LogtalkDefinitionProvider implements DefinitionProvider {
   private logger = getLogger();
@@ -79,16 +77,17 @@ export class LogtalkDefinitionProvider implements DefinitionProvider {
     }
     const def = path.join(dir, ".vscode_definition");
 
-    if (fs.existsSync(def)) {
-      const out = fs.readFileSync(def).toString();
-      await fsp.rm(def, { force: true });
+    try {
+      const content = await workspace.fs.readFile(Uri.file(def));
+      const out = content.toString();
+      await workspace.fs.delete(Uri.file(def), { useTrash: false });
       const match = out.match(/File:(.+);Line:(\d+)/);
       if (match) {
         let fileName = Utils.normalizeDoubleSlashPath(match[1]);
         const lineNum: number = parseInt(match[2]);
         location = new Location(Uri.file(fileName), new Position(lineNum - 1, 0));
       }
-    } else {
+    } catch (err) {
       this.logger.error('.vscode_definition file not found');
     }